+++ src/core/lib/transport/parsed_metadata.h
@@ -153,7 +153,7 @@ class ParsedMetadata {
   // Construct metadata from a string key, slice value pair.
   ParsedMetadata(Slice key, Slice value)
       : vtable_(ParsedMetadata::KeyValueVTable(key.as_string_view())),
-        transport_size_(key.size() + value.size() + 32) {
+        transport_size_((uint32_t)(key.size() + value.size() + 32)) {
     value_.pointer =
         new std::pair<Slice, Slice>(std::move(key), std::move(value));
   }
@@ -192,7 +192,7 @@ class ParsedMetadata {
     ParsedMetadata result;
     result.vtable_ = vtable_;
     result.value_ = value_;
-    result.transport_size_ = TransportSize(key().length(), value.length());
+    result.transport_size_ = TransportSize((uint32_t)key().length(), (uint32_t)value.length());
     vtable_->with_new_value(&value, on_error, &result);
     return result;
   }
